WebUI SD Size fix
fix for https://github.com/pr3y/Bruce/issues/904

+++ src/modules/others/webInterface.cpp
@@ -310,10 +310,10 @@ void configureWebServer() {
   });
   server->on("/systeminfo", HTTP_GET, []() {
     char response_body[250];
-    size_t LittleFSTotalBytes = LittleFS.totalBytes();
-    size_t LittleFSUsedBytes = LittleFS.usedBytes();
-    size_t SDTotalBytes = SD.totalBytes();
-    size_t SDUsedBytes = SD.usedBytes();
+    uint64_t LittleFSTotalBytes = LittleFS.totalBytes();
+    uint64_t LittleFSUsedBytes = LittleFS.usedBytes();
+    uint64_t SDTotalBytes = SD.totalBytes();
+    uint64_t SDUsedBytes = SD.usedBytes();
     sprintf(response_body,
       "{\"%s\":\"%s\",\"SD\":{\"%s\":\"%s\",\"%s\":\"%s\",\"%s\":\"%s\"},\"LittleFS\":{\"%s\":\"%s\",\"%s\":\"%s\",\"%s\":\"%s\"}}",
       "BRUCE_VERSION", BRUCE_VERSION,
